Update homework1.py
1. I created a variable "time_passed" and assigned it a value of  3800.
2. I developed the functions (get_hours_since_midnight, get_minutes & get_seconds),
making them count hours, minutes and seconds basing on integer division
and remainder division operators.
3. The "print" function outputs an informational string first,
telling the user what is going to be displayed, and after that consequentally
displays the result of previous calculations.
4. Also I added "str" function to the minutes inside the "print" function's arguments in order to
display the final result in time format. I'm sure there's a way to do that properly without changing
the type of the argument, but I don't know it yet.